requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0
watchdog>=3.0.0